+++ docs/data/api/popover-positioner.json
@@ -31,7 +31,6 @@
       "default": "5"
     },
     "container": { "type": { "name": "union", "description": "HTML element<br>&#124;&nbsp;func" } },
-    "finalFocus": { "type": { "name": "custom", "description": "ref" } },
     "hideWhenDetached": { "type": { "name": "bool" }, "default": "false" },
     "keepMounted": { "type": { "name": "bool" }, "default": "false" },
     "positionMethod": {
