+++ docs/data/translations/api-docs/popover-positioner/popover-positioner.json
@@ -21,9 +21,6 @@
       "description": "The padding between the popover element and the edges of the collision boundary to add whitespace between them to prevent them from touching."
     },
     "container": { "description": "The element the popover positioner element is appended to." },
-    "finalFocus": {
-      "description": "Determines an element to focus after the popover is closed. If not provided, the focus returns to the trigger."
-    },
     "hideWhenDetached": {
       "description": "Whether the popover element is hidden if it appears detached from its anchor element due to the anchor element being clipped (or hidden) from view."
     },
